pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dreamscribbles"
version = "0.1.0"
description = "DreamScribbles: Transform scribbles into stunning images using ControlNet + Stable Diffusion"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "DreamScribbles Maintainers" }
]
license = { text = "MIT" }
keywords = ["diffusers", "controlnet", "stable-diffusion", "gradio", "image-to-image", "ml", "ai-art"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

# Core runtime dependencies
# Note: Torch CPU builds are installed by default via pip; CUDA users can override in their env
# Apple Silicon uses MPS automatically with recent torch wheels
dependencies = [
  "diffusers==0.30.0",
  "transformers==4.37.2",
  "accelerate>=0.30.0",
  "torch>=2.0.0",
  "safetensors>=0.4.2",
  "pillow>=10.3.0",
  "numpy>=1.26.0,<2.0.0",
  "opencv-python-headless>=4.7,<4.10",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "python-dotenv>=1.0.1",
  "gradio==4.29.0",
  "huggingface-hub>=0.20.2,<1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "mypy>=1.10.0",
  "pre-commit>=3.7.0",
  "ruff>=0.6.9",
  "black>=24.8.0",
  "isort>=5.13.2",
  "types-Pillow>=10.2.0.20240822",
  "types-requests>=2.32.0.20240712"
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "SIM", "C90"]
ignore = []

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
warn_return_any = true
warn_unused_configs = true
pretty = true

[[tool.mypy.overrides]]
module = "gradio.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
